[feat] 좀비 스탯 보정치 적용
좀비 스탯 보정치 적용

+++ Zippers/Assets/Scripts/조경민/IResourceCollectable.cs
@@ -1,4 +1,4 @@
 public interface IResourceCollectable
 {
-    void CollectResource(ResourcesType type, int amount);
+    void CollectResource(ResourcesType type, float amount);
 }
