+++ Zippers/Assets/Scripts/조경민/Reward.cs
@@ -3,10 +3,10 @@ using UnityEngine;
 public class Reward : MonoBehaviour, IPoolable
 {
     private ResourcesType _type = ResourcesType.None; // 재화 타입
-    private int _amount; // 재화량
+    private float _amount; // 재화량
     private bool _isCollected;
 
-    public void Init(ResourcesType type, int amount)
+    public void Init(ResourcesType type, float amount)
     {
         _type = type;
         _amount = amount;
@@ -20,7 +20,7 @@ public class Reward : MonoBehaviour, IPoolable
     public void OnDespawn()
     {
         _type = ResourcesType.None;
-        _amount = 0;
+        _amount = 0f;
         _isCollected = false;
     }
